backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9; python_version < "3.14"
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pandas==2.2.2; python_version < "3.14"
scikit-learn==1.4.2; python_version < "3.14"
numpy==1.26.4; python_version < "3.14"
requests==2.31.0
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
gitpython==3.1.43
radon==6.0.1
